# gh-118218: Reuse return tuple in itertools.pairwise