--- a/Misc/NEWS.d/next/Library/2024-04-24-07-45-08.gh-issue-118218.m1OHbN.rst
+++ b/Misc/NEWS.d/next/Library/2024-04-24-07-45-08.gh-issue-118218.m1OHbN.rst
@@ -1,1 +1,1 @@
-Speed up :func:`itertools.pairwise` by up to 1.8x
+Speed up :func:`itertools.pairwise` in the common case by up to 1.8x.