@rhettinger 2024-04-25
Conceptually this seems reasonable to me.  Please do have someone else give it a thorough review (Serhiy would be a reasonable choice).