@hauntsaninja 2024-04-28
`PyTuple_New` + `PyTuple_SET_ITEM` (maybe 1.05x faster):
```
b1(1000)         min=27.0ms mean=27.3ms ± 0.1ms (50 repeats, 1000 loops)
b2(1000)         min=28.4ms mean=28.5ms ± 0.2ms (50 repeats, 1000 loops)
b3(1000)         min=25.2ms mean=25.7ms ± 0.3ms (50 repeats, 1000 loops)
```

Using this diff:
<details>

```diff
diff --git a/Modules/itertoolsmodule.c b/Modules/itertoolsmodule.c
index 6ee447ef6a..668c5a4f15 100644
--- a/Modules/itertoolsmodule.c
+++ b/Modules/itertoolsmodule.c
@@ -356,7 +356,11 @@ pairwise_next(pairwiseobject *po)
         return NULL;
     }
     /* Future optimization: Reuse the result tuple as we do in enumerate() */
-    result = PyTuple_Pack(2, old, new);
+    result = PyTuple_New(2);
+    if (result != NULL) {
+        PyTuple_SET_ITEM(result, 0, Py_NewRef(old));
+        PyTuple_SET_ITEM(result, 1, Py_NewRef(new));
+    }
     Py_XSETREF(po->old, new);
     Py_DECREF(old);
     return result;
```
</details>

Baseline:
```
b1(1000)         min=26.9ms mean=27.5ms ± 0.2ms (50 repeats, 1000 loops)
b2(1000)         min=29.5ms mean=29.7ms ± 0.3ms (50 repeats, 1000 loops)
b3(1000)         min=26.4ms mean=26.8ms ± 0.2ms (50 repeats, 1000 loops)
```

This PR:
```
b1(1000)         min=20.2ms mean=20.4ms ± 0.2ms (50 repeats, 1000 loops)
b2(1000)         min=16.9ms mean=17.0ms ± 0.0ms (50 repeats, 1000 loops)
b3(1000)         min=26.0ms mean=26.1ms ± 0.2ms (50 repeats, 1000 loops)
```